GuardianLens/backend/pyproject.toml
[tool.poetry]
name = "guardian_lens_backend"
version = "0.1.0"
description = "GuardianLens Backend API and Workers"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
  fastapi = "^0.111.0"
  uvicorn = {extras = ["standard"], version = "^0.29.0"}
  pydantic-settings = "^2.2.1"
  celery = "^5.3.6"
  redis = "^5.0.4"
    sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
    asyncpg = "^0.29.0"
    alembic = "^1.13.1"
    greenlet = "^3.0.3"
  passlib = {extras = ["bcrypt"], version = "^1.7.4"}
  python-jose = {extras = ["cryptography"], version = "^3.3.0"}
  python-multipart = "^0.0.9"
  google-auth-oauthlib = "^1.2.0"
  google-api-python-client = "^2.128.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.27.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"